+++ packages/svelte/formula/src/index.ts
@@ -3,15 +3,15 @@ import { createGroup } from './lib/group/group';
 import { Beaker, BeakerStores, Formula, FormulaError, FormulaOptions, FormulaStores, FormValues } from './types';
 import { createFormStores, createGroupStores } from './lib/shared/stores';
 
-export { FormulaError, FormValues, FormulaStores };
+export { Beaker, BeakerStores, Formula, FormulaError, FormulaOptions, FormulaStores, FormValues };
 
 /**
  * A global map of stores for elements with an `id` property and the `use` directive,
  * if no ID is used the store is not added
  * @type Map<string, FormulaStores>
  */
-export const formulaStores = new Map<string, FormulaStores>();
-export const beakerStores = new Map<string, BeakerStores>();
+export const formulaStores = new Map<string, FormulaStores<FormValues>>();
+export const beakerStores = new Map<string, BeakerStores<FormValues>>();
 
 /**
  * The `formula` function returns a form object that can be bound to any HTML
@@ -23,8 +23,8 @@ export const beakerStores = new Map<string, BeakerStores>();
  * @returns Formula object containing the current form, function to update or destroy
  * the form and all the stores available for the form
  */
-export function formula(options?: FormulaOptions): Formula {
-  const stores = createFormStores(options);
+export function formula<T extends FormValues>(options?: FormulaOptions): Formula<T> {
+  const stores = createFormStores<T>(options);
 
   const { create, update, destroy, reset } = createForm(stores, options, formulaStores);
 
@@ -39,18 +39,21 @@ export function formula(options?: FormulaOptions): Formula {
 }
 
 /**
- * The `beaker` function returns an object that allows for the creation of form groups within your main form
- * The methos returns a group
+ * The beaker function returns an instance of a group of elements and their stores, it also provides methods
+ * to set the group value store
+ *
  * @param options
+ *
+ * @returns Beaker object containing the form group and it's associated methods
  */
-export function beaker(options?: FormulaOptions): Beaker {
-  const stores = createGroupStores(options);
-  const { create, update, destroy, reset } = createGroup(stores, options, beakerStores);
+export function beaker<T extends FormValues>(options?: FormulaOptions): Beaker<T> {
+  const stores = createGroupStores<T>(options);
+  const { create, destroy, reset, forms } = createGroup<T>(stores, options, beakerStores as any);
   return {
     group: create,
-    update: update,
     destroy: destroy,
-    reset: reset,
+    forms,
+    reset,
     stores,
     ...stores,
     init: (items) => stores.formValues.set(items),
@@ -60,5 +63,6 @@ export function beaker(options?: FormulaOptions): Beaker {
         state.splice(index, 1);
         return state;
       }),
+    clear: () => stores.formValues.set([]),
   };
 }